Refactor to separate file to fix GitHub unused var build issue.
